Enhance the support for migration+deletion. (#338)
This PR overhaul the implementation of "signatures". We previously store signatures as

```
/signatures
     /sig -> object_id
```

Now we have

```
/signatures
    /instance_name
        /sig -> object_id
```

Such a change allows us to differ the replica for an object after migration, and do things correctly when deleting the migrated chunk or the original chunk after migration. The global object will be still kept complete and the deleted chunk could be freed as expected.

See also test case `test_migration_and_deletion` and `Note [Deleting global object and members]` for the detail about how we handle the use-ref in such cases.

Signed-off-by: Tao He <linzhu.ht@alibaba-inc.com>