Add CI test for spark and distributed hive. (#1705)
Add ci test for spark and distributed hive.

- Fixes #1655
- Fixes #1658
- Fixes #1471 

N.B.: Spark and Hive get stuck when conducting distributed testing in
different containers, whereas it does not happen when they are in the
same container. There may be a configuration issue somewhere, so for
now, we will temporarily put Spark and Hive together. This will not
affect the test results (the vineyard server for communicating with the
work node did receive requests to create and modify data).

Submitting the example program to yarn in the spark container worked fine.

Signed-off-by: vegetableysm <yuanshumin.ysm@alibaba-inc.com>